+++ src/com/bayviewglen/zork/items/Weapon.java
@@ -77,7 +77,7 @@ public class Weapon extends Item{
 	 */
 	public void ability(Entity e, Player p){
 		if (didHit()) {
-			e.setHealth(e.getHealth() - getDamage() * p.getDamageScaler() - criticalHit());
+			e.setHealth((int)(e.getHealth() - getDamage() * p.getDamageScaler() - criticalHit()));
 			System.out.println("You attack " + e.getName() + " with " + getName() + " (-" + getDamage() + ")");
 		}
 	}
